requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
flask-talisman
PyJWT
flask-jwt-extended
python-dotenv